правка ошибок в работе с url
